backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# ASR (Speech Recognition)
openai-whisper==20231117

# TTS (Text-to-Speech)
piper-tts==1.2.0

# Audio Processing
numpy==1.26.2
scipy==1.11.4
soundfile==0.12.1
librosa==0.10.1
pydub==0.25.1

# AI/LLM
langchain==0.0.340
langchain-community==0.0.1
ollama==0.1.6
openai==1.3.7

# Utilities
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1
websockets==12.0

# Logging & Monitoring
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
